# WebService/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "A Mini Product Ordering Service API.",
        "title": "Gin Mini Product Ordering Service",
        "contact": {},
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "paths": {
        "/Auth/login": {
            "post": {
                "description": "Login Form",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Login Form",
                "parameters": [
                    {
                        "type": "string",
                        "description": "email",
                        "name": "email",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "password",
                        "name": "password",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/Auth/register": {
            "post": {
                "description": "Register Form",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Register Form",
                "parameters": [
                    {
                        "description": "Register User",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.RegisterData"
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/Cart/AddItem": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Add Product to Cart",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Cart"
                ],
                "summary": "Add Product to Cart",
                "parameters": [
                    {
                        "description": "Add Product to Cart ",
                        "name": "product",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.CartItemSend"
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/Cart/EmptyCart/{uID}": {
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Remove All Product from Cart",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Cart"
                ],
                "summary": "Remove All Product from Cart",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Remove All Product from Cart ",
                        "name": "uID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/Cart/RemoveItem/{pID}": {
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Remove Product from Cart",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Cart"
                ],
                "summary": "Remove Product from Cart",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Remove Product from Cart ",
                        "name": "pID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/Cart/{uID}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Get All Product from Cart",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Cart"
                ],
                "summary": "Get All Product from Cart",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Get All Product from Cart ",
                        "name": "uID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
...
        "/Product/Search": {
            "post": {
                "description": "Search Product",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Product"
                ],
                "summary": "Search Product",
                "parameters": [
                    {
                        "type": "integer",
                        "format": "number",
                        "description": "page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "format": "number",
                        "description": "data per page",
                        "name": "recordPerPage",
                        "in": "query"
                    },
                    {
                        "description": "Search Product ",
                        "name": "product",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/models.SearchProduct"
                        }
                    }
                ],
                "responses": {}
            }
        },
...
        "/Product/{pid}": {
            "get": {
                "description": "Detail Product",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Product"
                ],
                "summary": "Detail Product",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Get Detail Product",
                        "name": "pid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
...
    },
    "definitions": {
...
        "models.CartItemSend": {
            "type": "object",
            "properties": {
                "productID": {
                    "type": "string"
                },
                "qty": {
                    "type": "integer"
                }
            }
        },
...
        "models.RegisterData": {
            "type": "object",
            "required": [
                "email",
                "password",
                "username"
            ],
            "properties": {
                "about": {
                    "type": "string"
                },
                "birthdate": {
                    "type": "integer"
                },
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string",
                    "minLength": 6
                },
                "phoneNumber": {
                    "type": "string"
                },
                "role": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "models.SearchProduct": {
            "type": "object",
            "properties": {
                "category": {
                    "type": "string"
                },
                "keyword": {
                    "type": "string",
                    "minLength": 3
                },
                "order": {
                    "type": "string"
                },
                "priceHigh": {
                    "type": "integer"
                },
                "priceLow": {
                    "type": "integer"
                },
                "rating": {
                    "type": "number"
                },
                "sortBy": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Token",
            "in": "header"
        }
    }
}